Added Django 1.8 xgettextmakemessages command
