+++ django_makemessages_xgettext/management/commands/xgettextmakemessages.py
@@ -1,21 +1,9 @@
-from optparse import make_option
-from django.core.management.commands import makemessages
+import django
 
+if django.get_version().startswith("1.8"):
+    from django_makemessages_xgettext import django18_xgettextmakemessages
+    Command = django18_xgettextmakemessages.Command
 
-class Command(makemessages.Command):
-
-    option_list = makemessages.Command.option_list + (
-        make_option('--xgettext', '-X',
-                    default=[], dest='xgettext', action='append',
-                    help='Passes arguments to the xgettext command (e.g. '
-                         '--add-location=file). Use multiple times to pass '
-                         'additional arguments.'
-                    ),
-    )
-
-    def handle_noargs(self, *args, **options):
-
-        for option in options["xgettext"]:
-            self.xgettext_options = self.xgettext_options[:] + [option]
-
-        super(Command, self).handle_noargs(*args, **options)
+else:
+    from django_makemessages_xgettext import django17_xgettextmakemessages
+    Command = django17_xgettextmakemessages.Command
