+++ test_project/test_project/locale/en/LC_MESSAGES/django.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-07-16 15:07+0000\n"
+"POT-Creation-Date: 2015-07-16 19:33+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -16,14 +16,14 @@ msgstr ""
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 
-#: test_app/views.py:6
+#: test_app/views.py
 msgid "Hello, world."
 msgstr ""
 
-#: test_project/settings.py:93
+#: test_project/settings.py
 msgid "English"
 msgstr ""
 
-#: test_project/settings.py:94
+#: test_project/settings.py
 msgid "French"
 msgstr ""
